fastapi-backend/requirements.txt
numpy>=1.26.4,<2
pandas>=2.0.0,<2.2.0
fastapi>=0.110.0,<0.116.0
lightfm>=1.17,<2
uvicorn>=0.29.0,<0.31.0
scikit-learn>=1.4.0,<1.6.0
transformers>=4.40.0,<4.50.0
torch>=2.0.0,<2.4.0
psycopg2-binary>=2.9.9,<3
huggingface_hub>=0.23.0
mab2rec>=1.3.1,<2
python-dotenv>=1.0.0
